Try to reduce size
